signal_system/pyproject.toml
[project]
name = "signal-system"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "akshare>=1.18.47",
    "alpaca-py>=0.43.2",
    "jupyter>=1.1.1",
    "lxml>=6.0.2",
    "matplotlib>=3.10.8",
    "nbformat>=5.10.4",
    "notebook>=7.5.5",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "yfinance>=1.2.0",
]
